requirements.txt
pygame>=2.1.0
pyserial>=3.5
PyQt5>=5.15.0
numpy>=1.21.0
Pillow>=8.3.0 